Add flake
`pipeline-manager` can now be run under Nix by first
entering a devshell with `nix develop`, and then running:

```
cargo run --no-default-features --bin=pipeline-manager -- --db-connection-string postgres://postgres@localhost:5432
```

The flag `no-default-features` is used to avoid using the library
`embedded_postgresql`, which works by downloading binaries outside of
the package manager, which it then unsuccessfully tries to run. Instead,
we setup a postgresql database manually and tell pipeline-manager to use
that one. (The database setup is not included in this repo; for now that
part has to be done manually)